backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
requests==2.31.0
PyGithub==2.1.1
arxiv==1.4.8
google-generativeai==0.3.2
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-dateutil==2.8.2
feedparser==6.0.10 